# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Couvoit API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/persons": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "List persons (admin)",
                "operationId": "personsIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Persons"
                ],
                "summary": "Complete my profile",
                "description": "Complete my profile.\n\nNow: authenticated user is User, profile is User->person.",
                "operationId": "personsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePersonRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/persons/{id}": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Get person by id",
                "operationId": "personsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Persons"
                ],
                "summary": "Delete person (soft)",
                "operationId": "personsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Persons"
                ],
                "summary": "Update person",
                "operationId": "personsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePersonRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/persons/{id}/trips-driver": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Trips as driver",
                "operationId": "personsTripsDriver",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/persons/{id}/trips-passenger": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Trips as passenger",
                "operationId": "personsTripsPassenger",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/person-role": {
            "patch": {
                "tags": [
                    "Persons"
                ],
                "summary": "Update user role (admin)",
                "operationId": "personsUpdateRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRolePersonRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePersonRequestPayload": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "John",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe",
                        "nullable": true
                    },
                    "firstname": {
                        "description": "Alias for first_name",
                        "type": "string",
                        "maxLength": 100,
                        "example": "John",
                        "nullable": true
                    },
                    "lastname": {
                        "description": "Alias for last_name",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 15,
                        "example": "+33600000000",
                        "nullable": true
                    },
                    "pseudo": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "jdoe",
                        "nullable": true
                    },
                    "car_id": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdatePersonRequestPayload": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "John",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe",
                        "nullable": true
                    },
                    "firstname": {
                        "description": "Alias for first_name",
                        "type": "string",
                        "maxLength": 100,
                        "example": "John",
                        "nullable": true
                    },
                    "lastname": {
                        "description": "Alias for last_name",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 15,
                        "example": "+33600000000",
                        "nullable": true
                    },
                    "pseudo": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "jdoe",
                        "nullable": true
                    },
                    "car_id": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "DELETED"
                        ],
                        "example": "ACTIVE",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateRolePersonRequestPayload": {
                "required": [
                    "person_id",
                    "role_id"
                ],
                "properties": {
                    "person_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}